pyproject.toml
[project]
name = "lifemetrics"
version = "1.0.0"
description = "Health tracking backend"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "django>=5.0,<5.1",
    "django-ninja>=1.3.0",
    "uvicorn[standard]>=0.27.0",
    "gunicorn>=21.2.0",
    "psycopg[binary,pool]>=3.1.18",
    "asgiref>=3.7.2",
    "redis[hiredis]>=5.0.1",
    "django-redis>=5.4.0",
    "celery[redis]>=5.3.6",
    "django-celery-beat>=2.6.0",
    "django-celery-results>=2.5.1",
    "aiogram>=3.4.1",
    "aiohttp>=3.9.1",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
    "argon2-cffi>=23.1.0",
    "django-cors-headers>=4.3.1",
    "sentry-sdk>=1.39.2",
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",
    "python-dotenv>=1.0.0",
    "django-environ>=0.11.2",
    "orjson>=3.9.10",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "httptools>=0.6.1",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-django>=4.7.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "config*"]

[tool.ruff]
target-version = "py311"
line-length = 120
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["apps", "config"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*", "*/tests/*"]
